requirements.txt
colorama
python-dotenv
openai
streamlit